lab4/requirements.txt
transformers
aiogram
python-dotenv
accelerate
bitsandbytes
optimum
numpy